LayoutTests/http/tests/security/xss-DENIED-window-index-assign-expected.txt

CONSOLE MESSAGE: undefined
CONSOLE MESSAGE: TypeError: parent[0].f is not a function. (In 'parent[0].f()', 'parent[0].f' is undefined)

This test passes if the access is forbidden.
